refactor(test): extract assertRealTimestamp helper and update pm-handoff tests #654
